+++ js/connection.ts
@@ -77,11 +77,12 @@ class ServerConnection {
         const self = this;
         try {
             this._connectTimeoutHandler = setTimeout(() => {
+                console.log("Connect timeout triggered!");
                 this.disconnect();
                 this._client.handleDisconnect(DisconnectReason.CONNECT_FAILURE);
             }, timeout);
             let sockCpy;
-            this._socket = (sockCpy = new WebSocket('wss:' + address.host + ":" + address.port));
+            this._socket = (sockCpy = new WebSocket('wss://' + address.host + ":" + address.port));
             clearTimeout(this._connectTimeoutHandler);
             this._connectTimeoutHandler = null;
             if(this._socket != sockCpy) return; //Connect timeouted
@@ -120,6 +121,7 @@ class ServerConnection {
 
     updateConnectionState(state: ConnectionState) {
         this._connectionState = state;
+        this._client.controlBar.update_connection_state();
     }
 
     disconnect() : boolean {
@@ -273,9 +275,11 @@ class HandshakeHandler {
     readonly identity: Identity;
     readonly name?: string;
     private connection: ServerConnection;
+    server_password: string;
 
-    constructor(identity: Identity, name?: string) {
+    constructor(identity: Identity, name?: string, password?: string) {
         this.identity = identity;
+        this.server_password = password;
         this.name = name;
     }
 
@@ -293,11 +297,17 @@ class HandshakeHandler {
             data.publicKey = (this.identity as TeamSpeakIdentity).publicKey();
         } else if(this.identity.type() == IdentitifyType.TEAFORO) {
             data.data = (this.identity as TeaForumIdentity).identityDataJson;
+        } else if(this.identity.type() == IdentitifyType.NICKNAME) {
+            data["client_nickname"] = this.identity.name();
         }
 
         this.connection.sendCommand("handshakebegin", data).catch(error => {
             console.log(error);
             //TODO here
+        }).then(() => {
+            if(this.identity.type() == IdentitifyType.NICKNAME) {
+                this.handshake_finished();
+            }
         });
     }
 
@@ -307,19 +317,37 @@ class HandshakeHandler {
             proof = (this.identity as TeamSpeakIdentity).signMessage(json[0]["message"]);
         } else if(this.identity.type() == IdentitifyType.TEAFORO) {
             proof = (this.identity as TeaForumIdentity).identitySign;
+        } else if(this.identity.type() == IdentitifyType.NICKNAME) {
+            //FIXME handle error this should never happen!
         }
         this.connection.sendCommand("handshakeindentityproof", {proof: proof}).then(() => {
-            this.connection.sendCommand("clientinit", {
-                //TODO variables!
-                client_nickname: this.name ? this.name : this.identity.name(),
-                client_platform: navigator.platform,
-                client_version: navigator.userAgent,
-                //client_browser_engine: navigator.product
-            });
+           this.handshake_finished();
         }).catch(error => {
             console.error("Got login error");
             console.log(error);
         }); //TODO handle error
+    }
+
+    private handshake_finished() {
+        this.connection.sendCommand("clientinit", {
+            //TODO variables!
+            client_nickname: this.name ? this.name : this.identity.name(),
+            client_platform: navigator.platform,
+            client_version: navigator.userAgent,
+
+            client_server_password: this.server_password
+            //client_browser_engine: navigator.product
+        }).catch(error => {
+            this.connection.disconnect();
+            if(error instanceof CommandResult) {
+                if(error.id == 1028) {
+                    this.connection._client.handleDisconnect(DisconnectReason.SERVER_REQUIRES_PASSWORD);
+                } else {
+
+                    this.connection._client.handleDisconnect(DisconnectReason.CLIENT_KICKED, error);
+                }
+            }
+        });
     }
 }