+++ js/voice/VoiceHandler.ts
@@ -2,6 +2,8 @@
 /// <reference path="../codec/Codec.ts" />
 /// <reference path="VoiceRecorder.ts" />
 
+import WarningListener = NodeJS.WarningListener;
+
 class CodecPoolEntry {
     instance: BasicCodec;
     owner: number;
@@ -101,12 +103,29 @@ class CodecPool {
     }
 }
 
+enum VoiceConnectionType {
+    JS_ENCODE,
+    NATIVE_ENCODE
+}
+
+/* funny fact that typescript dosn't find this */
+interface RTCPeerConnection {
+    addStream(stream: MediaStream): void;
+    getLocalStreams(): MediaStream[];
+    getStreamById(streamId: string): MediaStream | null;
+    removeStream(stream: MediaStream): void;
+    createOffer(successCallback?: RTCSessionDescriptionCallback, failureCallback?: RTCPeerConnectionErrorCallback, options?: RTCOfferOptions): Promise<RTCSessionDescription>;
+}
+
 class VoiceConnection {
     client: TSClient;
     rtcPeerConnection: RTCPeerConnection;
     dataChannel: RTCDataChannel;
 
     voiceRecorder: VoiceRecorder;
+    private _type: VoiceConnectionType = VoiceConnectionType.NATIVE_ENCODE;
+
+    local_audio_stream: any;
 
     private codec_pool: CodecPool[] = [
         new CodecPool(this,0,"Spex A", undefined), //Spex
@@ -123,25 +142,94 @@ class VoiceConnection {
 
     constructor(client) {
         this.client = client;
+        this._type = settings.static_global("voice_connection_type", this._type);
         this.voiceRecorder = new VoiceRecorder(this);
-        this.voiceRecorder.on_data = this.handleVoiceData.bind(this);
         this.voiceRecorder.on_end = this.handleVoiceEnded.bind(this);
+        this.voiceRecorder.on_start = this.handleVoiceStarted.bind(this);
         this.voiceRecorder.reinitialiseVAD();
 
         AudioController.on_initialized(() => {
+            log.info(LogCategory.VOICE, "Initializing voice handler after AudioController has been initialized!");
             this.codec_pool[4].initialize(2);
             this.codec_pool[5].initialize(2);
+
+            if(this.type == VoiceConnectionType.NATIVE_ENCODE)
+                this.setup_native();
+            else
+                this.setup_js();
         });
 
         this.send_task = setInterval(this.sendNextVoicePacket.bind(this), 20);
+    }
+
+    native_encoding_supported() : boolean {
+        if(!(window.webkitAudioContext || window.AudioContext || {prototype: {}} as typeof AudioContext).prototype.createMediaStreamDestination) return false; //Required, but not available within edge
+        return true;
+    }
+
+    javascript_encoding_supported() : boolean {
+        if(!(window.RTCPeerConnection || {prototype: {}} as typeof RTCPeerConnection).prototype.createDataChannel) return false;
+        return true;
+    }
+
+    current_encoding_supported() : boolean {
+        switch (this._type) {
+            case VoiceConnectionType.JS_ENCODE:
+                return this.javascript_encoding_supported();
+            case VoiceConnectionType.NATIVE_ENCODE:
+                return this.native_encoding_supported();
+        }
+        return false;
+    }
+
+    private setup_native() {
+        log.info(LogCategory.VOICE, "Setting up native voice stream!");
+        if(!this.native_encoding_supported()) {
+            log.warn(LogCategory.VOICE, "Native codec isnt supported!");
+            return;
+        }
+
+        this.voiceRecorder.on_data = undefined;
+
+        let stream =  this.voiceRecorder.get_output_stream();
+        stream.disconnect();
+
+        if(!this.local_audio_stream)
+            this.local_audio_stream = AudioController.globalContext.createMediaStreamDestination();
+        stream.connect(this.local_audio_stream);
+    }
+
+    private setup_js() {
+        if(!this.javascript_encoding_supported()) return;
+
+        this.voiceRecorder.on_data = this.handleVoiceData.bind(this);
+    }
+
+    get type() : VoiceConnectionType { return this._type; }
+    set type(target: VoiceConnectionType) {
+        if(target == this.type) return;
+        this._type = target;
+
+        if(this.type == VoiceConnectionType.NATIVE_ENCODE)
+            this.setup_native();
+        else
+            this.setup_js();
+        this.createSession();
     }
 
     codecSupported(type: number) : boolean {
         return this.codec_pool.length > type && this.codec_pool[type].supported();
     }
 
-    voiceSupported() : boolean {
+    voice_playback_support() : boolean {
         return this.dataChannel && this.dataChannel.readyState == "open";
+    }
+
+    voice_send_support() : boolean {
+        if(this.type == VoiceConnectionType.NATIVE_ENCODE)
+            return this.native_encoding_supported() && this.rtcPeerConnection.getLocalStreams().length > 0;
+        else
+            return this.voice_playback_support();
     }
 
     private voice_send_queue: {data: Uint8Array, codec: number}[] = [];
@@ -178,7 +266,13 @@ class VoiceConnection {
 
 
     createSession() {
+        if(!this.current_encoding_supported()) return false;
+
+        if(this.rtcPeerConnection) {
+            this.dropSession();
+        }
         this._ice_use_cache = true;
+
 
         let config: RTCConfiguration = {};
         config.iceServers = [];
@@ -192,10 +286,15 @@ class VoiceConnection {
         this.dataChannel.binaryType = "arraybuffer";
 
         let sdpConstraints : RTCOfferOptions = {};
-        sdpConstraints.offerToReceiveAudio = 0;
-        sdpConstraints.offerToReceiveVideo = 0;
+        sdpConstraints.offerToReceiveAudio = this._type == VoiceConnectionType.NATIVE_ENCODE;
+        sdpConstraints.offerToReceiveVideo = false;
 
         this.rtcPeerConnection.onicecandidate = this.onIceCandidate.bind(this);
+
+        if(this.local_audio_stream) { //May a typecheck?
+            this.rtcPeerConnection.addStream(this.local_audio_stream.stream);
+            console.log("Adding stream (%o)!", this.local_audio_stream.stream);
+        }
         this.rtcPeerConnection.createOffer(this.onOfferCreated.bind(this), () => {
             console.error("Could not create ice offer!");
         }, sdpConstraints);
@@ -212,15 +311,21 @@ class VoiceConnection {
     handleControlPacket(json) {
         if(json["request"] === "answer") {
             console.log("Set remote sdp! (%o)", json["msg"]);
-            this.rtcPeerConnection.setRemoteDescription(new RTCSessionDescription(json["msg"]));
+            this.rtcPeerConnection.setRemoteDescription(new RTCSessionDescription(json["msg"])).catch(error => {
+                console.log("Failed to apply remote description: %o", error); //FIXME error handling!
+            });
             this._ice_use_cache = false;
             for(let msg of this._ice_cache) {
-                this.rtcPeerConnection.addIceCandidate(new RTCIceCandidate(msg));
+                this.rtcPeerConnection.addIceCandidate(new RTCIceCandidate(msg)).catch(error => {
+                    console.log("Failed to add remote cached ice candidate %s: %o", msg, error);
+                });
             }
         } else if(json["request"] === "ice") {
             if(!this._ice_use_cache) {
                 console.log("Add remote ice! (%s | %o)", json["msg"], json);
-                this.rtcPeerConnection.addIceCandidate(new RTCIceCandidate(json["msg"]));
+                this.rtcPeerConnection.addIceCandidate(new RTCIceCandidate(json["msg"])).catch(error => {
+                    console.log("Failed to add remote ice candidate %s: %o", json["msg"], error);
+                });
             } else {
                 console.log("Cache remote ice! (%s | %o)", json["msg"], json);
                 this._ice_cache.push(json["msg"]);
@@ -241,18 +346,28 @@ class VoiceConnection {
     onIceCandidate(event) {
         console.log("Got ice candidate! Event:");
         console.log(event);
-        if (event && event.candidate) {
-            this.client.serverConnection.sendData(JSON.stringify({
-                type: 'WebRTC',
-                request: "ice",
-                msg: event.candidate,
-            }));
+        if (event) {
+            if(event.candidate)
+                this.client.serverConnection.sendData(JSON.stringify({
+                    type: 'WebRTC',
+                    request: "ice",
+                    msg: event.candidate,
+                }));
+            else {
+                this.client.serverConnection.sendData(JSON.stringify({
+                    type: 'WebRTC',
+                    request: "ice_finish"
+                }));
+            }
         }
     }
 
     onOfferCreated(localSession) {
         console.log("Offer created and accepted");
-        this.rtcPeerConnection.setLocalDescription(localSession);
+        this.rtcPeerConnection.setLocalDescription(localSession).catch(error => {
+            console.log("Failed to apply local description: %o", error);
+            //FIXME error handling
+        });
 
         console.log("Send offer: %o", localSession);
         this.client.serverConnection.sendData(JSON.stringify({type: 'WebRTC', request: "create", msg: localSession}));
@@ -318,12 +433,21 @@ class VoiceConnection {
     }
 
     private handleVoiceEnded() {
+        if(this.client && this.client.getClient())
+            this.client.getClient().speaking = false;
+
         if(!this.voiceRecorder) return;
         if(!this.client.connected) return;
+        console.log("Local voice ended");
 
-        console.log("Voice ended");
-        this.client.getClient().speaking = false;
         if(this.dataChannel)
-            this.sendVoicePacket(new Uint8Array(0), 4); //TODO Use channel codec!
+            this.sendVoicePacket(new Uint8Array(0), 5); //TODO Use channel codec!
+    }
+
+    private handleVoiceStarted() {
+        console.log("Local voice started");
+
+        if(this.client && this.client.getClient())
+            this.client.getClient().speaking = true;
     }
 }