+++ ChangeLog.md
@@ -1,4 +1,21 @@
 # Changelog:
+* **26.09.18**:
+    - Added Safari support
+    
+* **25.09.18**:
+    - Added support for token use
+    - Added support for away messages
+    * Fixed away display within information bar
+    - Capturing last given address and nickname within connect modal
+    * Using random password field ids for server connect modal
+    + Improved forum not authenticated message within connect modal
+    - Added partitional MS Edge support
+    * Fixed chat new line indeed
+    
+* **24.09.18**:
+    - Added server passwords within login modal
+    - Fixed native encoding for opera and firefox
+    - Made the default music bot group configurable
 
 * **0.0.1-Alpha**:  
 - First public upload of the Web client